[clad] Initial support of custom derivatives in clad for some of TMath functions
Clad allows to specify a custom derivative for any function. The derivative has to be defined in the `custom_derivatives` namespace and will be used
whenever the derivative of the original function is required (instead of proceeding to automatically derive the function).

For example, instead of deriving a practival implementation of `sin` (which is likely to be non-differentiable due to implementation details), we can
define
```
namespace custom_derivatives {
  double sin_darg0(double x) { return cos(x); }
}
```
As a result, `cos` will be used directly whenever the derivative of `sin` is required.

This commit adds custom derivatives for some of the mathematical functions in the `TMath` namespace.