[Exp PyROOT] No need to explicitly pythonize TNtuple and TNtupleD
TNtuple and TNtupleD will inherit from TTree all the pythonizations
added in the TTree pythonizor. Pythonizations on a superclass are
inherited by a subclass if the superclass adds a new method or
pythonizes an existing method that is not redefined in C++ in
the subclass. If the subclass redefines the method in C++, a
reinjection of the pythonization needs to be done in the subclass.